+++ src/comparer/boring.rs
@@ -18,8 +18,6 @@ impl Comparer for Boring {
             //l*a*b* squared distance should be precise enough
             let fdiff = s.squared_distance(d); //max: 30000
 
-            //let fdiff = ( s.l.max(d.l) - s.l.min(d.l) ) + ( s.a.max(d.a) - s.a.min(d.a) ) + ( s.b.max(d.b) - s.b.min(d.b) );
-
             let amul = *(sa.min(da)) as f32;// / 65025.0;
             sum += (fdiff*amul) as u64;
             //sum=sum.checked_add(((fdiff*amul)*(fdiff*amul)) as u64).unwrap();
@@ -27,12 +25,6 @@ impl Comparer for Boring {
 
         sum
     }
-    /*let fdiff = s.squared_distance(d); //max: 30000
-
-            //let fdiff = ( s.l.max(d.l) - s.l.min(d.l) ) + ( s.a.max(d.a) - s.a.min(d.a) ) + ( s.b.max(d.b) - s.b.min(d.b) );
-
-            let amul = *(sa.min(da)) as f32 * 65025.0;//65025.0;
-            sum = sum.max((fdiff*amul) as u64);*/
 
     fn pre_parse(i: DynamicImage, dest: (u32,u32), scale: FilterType) -> Self::DestImage {
         let i = Self::pre_parse2(i, dest, scale);