+++ src/composing/mod.rs
@@ -1,11 +1,8 @@
-use image::Rgba;
 use std::path::Path;
 use std::collections::HashMap;
 use std::sync::{RwLock, Arc};
 use crate::util::transfer;
 use image::RgbaImage;
-use crate::tiles::src_tile::SrcTile;
-use crate::tiles::dest_tile::DestTile;
 use crate::comparer::Comparer;
 use crate::meta::ArcMeta;
 use crate::util::RefClonable;
@@ -13,10 +10,8 @@ use std::fs::read;
 use rayon::iter::IntoParallelRefIterator;
 use rayon::iter::ParallelIterator;
 
-//pub mod decode;
-
 pub fn compose_walls<C: Comparer + Send + 'static>(mappings: &[Option<u32>], tile_files: &[Arc<Path>], cscale: u32, meta: ArcMeta<C>) -> RgbaImage where C::DestImage: Send + Sync {
-    //this inefficient hashmap vec thing is still faster than processing pixel masses
+    // collect the wall target per tile
     let mut map: HashMap<Arc<Path>,Vec<usize>> = HashMap::with_capacity(mappings.len());
 
     for (i,t) in mappings.iter().enumerate() {
@@ -49,21 +44,14 @@ pub fn compose_walls<C: Comparer + Send + 'static>(mappings: &[Option<u32>], til
         });
     };
 
-    /*for i in 0..(tcw*tch) {
-        let ox = (i as u32 % tcw) * tw; 
-        let oy = (i as u32 / tcw) * th;
-
-        out.put_pixel(ox,oy, Rgba([(i%255) as u8,0,0,255]));
-    }*/
-
     out
 }
 
 pub fn load<C: Comparer>(f: &Arc<Path>, cscale: u32, m: &ArcMeta<C>) -> RgbaImage {
     println!("\t{}",f.to_string_lossy());
-    let mem = read(&*f).expect("Image failed to read in second pass sowwy");
+    let mem = read(&*f).expect("Failed to read Image file");
 
-    let img = image::load_from_memory(&mem[..]).expect("Image suddenly broken is second pass sowwy");
+    let img = image::load_from_memory(&mem[..]).expect("Failed to decode Image file");
 
     drop(mem);