+++ src/decoding/mod.rs
@@ -1,14 +1,12 @@
 use std::error::Error;
-use crate::tiles::DestId;
 use async_std::fs::read;
 use crate::meta::ArcMeta;
 use async_std::task::block_on;
 use async_std::task::spawn;
-use crate::tiles::src_tile::SrcTile;
 use std::sync::Arc;
 use std::path::Path;
 use crate::comparer::Comparer;
-use crate::{puzzler::Match, util::RefClonable};
+use crate::{meta::Match, util::RefClonable};
 use itertools::Itertools;
 
 pub mod files;
@@ -23,12 +21,6 @@ pub async fn decode_and_compare<C: Comparer>(f: Arc<Path>, m: ArcMeta<C>, tile_i
     drop(mem);
 
     let iimg = C::pre_parse(img, m.tile_size, m.scale);
-
-    /*let mut compares: Vec<DestId> = m.walls_parsed.iter().enumerate()
-        .map(|(i,t)| DestId{id: i, diff: C::compare(&iimg, t, m.tile_size)} )
-        .collect();*/
-
-    //compares.sort_unstable_by_key(|k| k.diff);
 
     let mut dest = Vec::with_capacity(m.walls_parsed.len());