+++ src/opts.rs
@@ -1,4 +1,0 @@
-pub struct Opts {
-    tile_w: usize,
-    tile_h: usize,
-}