+++ src/puzzler/alltoall.rs
@@ -1,6 +1,5 @@
 use super::*;
 use std::collections::HashSet;
-use crate::tiles::*;
 
 pub struct AllToAll;