+++ src/puzzler/mod.rs
@@ -1,21 +1,12 @@
-use crate::tiles::src_tile::SrcTile;
-use crate::tiles::dest_tile::DestTile;
-
 pub mod simplemal;
 pub mod alltoall;
 
 use simplemal::*;
 use alltoall::*;
+use crate::meta::Match;
 
 pub trait Puzzler {
     fn puzzle(matches: Vec<Match>, wall: &mut [Option<u32>], tiles: usize, noise: usize);
-}
-
-#[derive(Clone,Copy)]
-pub struct Match {
-    pub tile: u32,
-    pub wall: u32,
-    pub diff: u64,
 }
 
 pub static PUZZLERS: &'static str = "simple/alltoall";