+++ src/tiles/dest_tile.rs
@@ -1,8 +1,0 @@
-use crate::tiles::SrcId;
-use image::DynamicImage;
-use crate::comparer::Comparer;
-
-pub struct DestTile {
-    pub src_matches: Vec<SrcId>,
-    pub linked: Option<SrcId>,
-}