+++ src/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - feat: all-to-all mapper
   - improved mapper which tries to match all source tiles
+- feat: compose scale factor
+  - multiplier for composing image/tile size
 - compose: multithreaded image decoding
 - optimized memory usage
 - reworked mapper to use less vecs